Use filelists to not remove exec permissions from eln files
We need to remove the exec permission on the eln files to prevent them from
having the same build-id and thereby failing the debuginfo extraction, while at
the same time they must be executable for emacs to use them.
We use filelists for that where we use attr to set their executable bits *after*
the build.